+++ src/utils/settings.py
@@ -262,7 +262,7 @@ DESCRIPTION_FULL = "Automated All-in-One OS Command Injection Exploitation Tool"
 DESCRIPTION = "The command injection exploiter"
 AUTHOR  = "Anastasios Stasinopoulos"
 VERSION_NUM = "4.1"
-REVISION = "31"
+REVISION = "32"
 STABLE_RELEASE = False
 VERSION = "v"
 if STABLE_RELEASE: